Implement support of CXX math functions for AMD
Add libcxx support for AMD in a similar fashion as 1fe92c5
Include code for the AMDGCN target by modifying macros.

Reuse the "-fbundle-no-offload-arch" for AMD.

This is a WP and will fail on all targets that are not gfx940.
The reason is that libdevice is built for gfx940 only.
It will not work on any other targets due to mismatching metadata.